Create a generic widget to view documentation
This creates DocumentationWidget which will display the Token
documentation and the Operator documentation in a QTabWidget.

It has a hook in `showEvent` that will regenerate the documentation,
however a manual refresh can be trigger via
`DocumentationWidget.refreshDocumentation`

**Warning** This takes the html text from the sept.Operator and sept.Token docstrings and directly loads it into a QWebView.
This is a known security risk if you are loading Tokens or Operators from an untrusted source.