Create a live preview input widget for sept
TemplateInputWidget can be used to interactively create `sept.Template` objects and check their validity.

All that TemplateInputWidget needs is a valid `sept.PathTemplateParser` instance, however for styling in your GUI application, you can optionally pass an `error_colour` that will drive the highlighting colour when errors occur.

*Interactivity Timeouts*
You may also wish to define how long the timeout occurs between the last character input and our error highlighting occurs.
By default this is set at 1250ms but you can override this by passing a value to `timeout` during instantiation.

*Error handling*
When using this in a larger GUI application, you may want to have a centralized place for displaying errors, if that is the case, you can send errors to the `recieve_error` slot.
However, to ensure it visualizes correctly, you will want to ensure your error class has "location" and "length" attributes on it that can be used to display the highlighting.